Fixing PHP Warning on Accessing "children" Property
I fixed an issue that caused a PHP warning when trying to access the "children" property on a boolean value. The error occurred due to an attempt to access a method on a variable that in some cases could be a boolean, instead of the expected object. For example, local_aiquestions couldn't work properly without this code change. So the check was added to ensure that $node->children is an object before attempting to access the type() method. This avoids the mentioned error and improves the stability and reliability of the code.